requirements.txt
hera>=5.0.0
nbformat>=5.0.0
GitPython>=3.1.0
python-dotenv>=1.0.0
